+++ src/thinger/config.h
@@ -1,5 +1,8 @@
 #include <nlohmann/json.hpp>
 
+#include <boost/random/random_device.hpp>
+#include <boost/random/uniform_int_distribution.hpp>
+
 #include <map>
 #include <iostream>
 #include <fstream>
@@ -32,14 +35,13 @@ namespace thinger::monitor::utils {
 
         const std::string CHARACTERS = "0123456789abcdefghijklmnopqrstuvwxyz!@#$%^&*()ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
-        std::random_device random_device;
-        std::mt19937 generator(random_device());
-        std::uniform_int_distribution distribution(0, static_cast<int>(CHARACTERS.size()) - 1);
+        boost::random::random_device rng;
+        boost::random::uniform_int_distribution index_dist(0, static_cast<int>(CHARACTERS.size()) - 1);
 
         std::string random_string;
 
-        for (std::size_t i = 0; i < length; ++i) {
-            random_string += CHARACTERS[distribution(generator)];
+        for(std::size_t i = 0; i < length; ++i) {
+            random_string += CHARACTERS[index_dist(rng)];
         }
 
         return random_string;