+++ src/thinger/utils/date.h
@@ -17,7 +17,7 @@ public:
         date = std::chrono::system_clock::to_time_t(now);
     }
 
-    std::string to_iso8601(const char del = '-', const bool extended = false, const std::string timezone = "local") {
+    std::string to_iso8601(const char del = '-', const bool extended = false, const std::string_view& tmz = "local") {
 
         std::string format =
             std::string("%Y").append(std::string(1,del))+
@@ -31,9 +31,9 @@ public:
                 std::string("%m").append(std::string(1,del))+
                 "%dT%H%M%SZ";
 
-        format.erase(remove(format.begin(), format.end(), '\0'), format.end()); //remove '\0' from string
+        std::erase(format, '\0'); //remove '\0' from string
 
-        if (timezone == "utc" || timezone == "gmt") {
+        if (tmz == "utc" || tmz == "gmt") {
             gmtime_r(&date, &time); // Compliant
             ss << std::put_time(&time, format.c_str());
         } else { // (timezone == "local")